requirements.txt
python-dotenv
urllib3==1.25.8